rename the xml classes as well, since the annotation provides compatibility with gml.
