+++ source/imaer-gml/src/main/java/nl/overheid/aerius/gml/v4_0/collection/FeatureMemberProperty.java
@@ -26,7 +26,7 @@ import nl.overheid.aerius.gml.v4_0.base.FeatureMemberImpl;
 import nl.overheid.aerius.gml.v4_0.building.Building;
 import nl.overheid.aerius.gml.v4_0.measure.SRM1RoadMeasureArea;
 import nl.overheid.aerius.gml.v4_0.result.CustomCalculationPoint;
-import nl.overheid.aerius.gml.v4_0.result.NSLCalculationPoint;
+import nl.overheid.aerius.gml.v4_0.result.CIMLKCalculationPoint;
 import nl.overheid.aerius.gml.v4_0.result.ReceptorPoint;
 import nl.overheid.aerius.gml.v4_0.source.EmissionSource;
 import nl.overheid.aerius.gml.v4_0.source.farmland.FarmlandEmissionSource;
@@ -79,7 +79,7 @@ public class FeatureMemberProperty extends AbstractProperty<FeatureMemberImpl> {
       @XmlElement(namespace = CalculatorSchema.NAMESPACE, name = "MooringInlandShippingEmissionSource", type = MooringInlandShippingEmissionSource.class),
       @XmlElement(namespace = CalculatorSchema.NAMESPACE, name = "CalculationPoint", type = CustomCalculationPoint.class),
       @XmlElement(namespace = CalculatorSchema.NAMESPACE, name = "ReceptorPoint", type = ReceptorPoint.class),
-      @XmlElement(namespace = CalculatorSchema.NAMESPACE, name = "NSLCalculationPoint", type = NSLCalculationPoint.class),
+      @XmlElement(namespace = CalculatorSchema.NAMESPACE, name = "NSLCalculationPoint", type = CIMLKCalculationPoint.class),
       @XmlElement(namespace = CalculatorSchema.NAMESPACE, name = "SRM1RoadDispersionLine", type = SRM1RoadDispersionLine.class),
       @XmlElement(namespace = CalculatorSchema.NAMESPACE, name = "SRM1RoadMeasureArea", type = SRM1RoadMeasureArea.class),
       @XmlElement(namespace = CalculatorSchema.NAMESPACE, name = "Building", type = Building.class)
