+++ source/imaer-gml/src/main/java/nl/overheid/aerius/gml/v5_0/togml/Result2GML.java
@@ -31,7 +31,7 @@ import nl.overheid.aerius.gml.v5_0.result.AbstractCalculationPoint;
 import nl.overheid.aerius.gml.v5_0.result.CalculationPointCorrection;
 import nl.overheid.aerius.gml.v5_0.result.CalculationPointCorrectionProperty;
 import nl.overheid.aerius.gml.v5_0.result.CustomCalculationPoint;
-import nl.overheid.aerius.gml.v5_0.result.NSLCalculationPoint;
+import nl.overheid.aerius.gml.v5_0.result.CIMLKCalculationPoint;
 import nl.overheid.aerius.gml.v5_0.result.ReceptorPoint;
 import nl.overheid.aerius.gml.v5_0.result.Result;
 import nl.overheid.aerius.gml.v5_0.result.ResultProperty;
@@ -44,7 +44,6 @@ import nl.overheid.aerius.shared.domain.result.EmissionResultType;
 import nl.overheid.aerius.shared.domain.v2.geojson.Geometry;
 import nl.overheid.aerius.shared.domain.v2.geojson.Point;
 import nl.overheid.aerius.shared.domain.v2.cimlk.CIMLKCorrection;
-import nl.overheid.aerius.shared.domain.v2.point.CIMLKCalculationPoint;
 import nl.overheid.aerius.shared.domain.v2.point.CalculationPoint;
 import nl.overheid.aerius.shared.domain.v2.point.CalculationPointFeature;
 import nl.overheid.aerius.shared.exception.AeriusException;
@@ -94,9 +93,9 @@ final class Result2GML {
 
   private AbstractCalculationPoint determineSpecificType(final CalculationPoint aeriusPoint, final Point point) throws AeriusException {
     final AbstractCalculationPoint returnPoint;
-    if (aeriusPoint instanceof CIMLKCalculationPoint) {
+    if (aeriusPoint instanceof nl.overheid.aerius.shared.domain.v2.point.CIMLKCalculationPoint) {
       //treat as a custom calculation point with added properties
-      returnPoint = fromNslCalculationPoint((CIMLKCalculationPoint) aeriusPoint);
+      returnPoint = fromAeriusCalculationPoint((nl.overheid.aerius.shared.domain.v2.point.CIMLKCalculationPoint) aeriusPoint);
     } else if (aeriusPoint instanceof nl.overheid.aerius.shared.domain.v2.point.CustomCalculationPoint) {
       //treat as a custom calculation point
       //The proper representation of a custompoint would be a circle with a surface of 1ha.
@@ -111,11 +110,11 @@ final class Result2GML {
     return returnPoint;
   }
 
-  private AbstractCalculationPoint fromNslCalculationPoint(final CIMLKCalculationPoint aeriusNSLPoint) {
-    final NSLCalculationPoint nslPoint = new NSLCalculationPoint();
-    nslPoint.setRejectionGrounds(aeriusNSLPoint.getRejectionGrounds());
-    nslPoint.setMonitorSubstance(aeriusNSLPoint.getMonitorSubstance());
-    return nslPoint;
+  private AbstractCalculationPoint fromAeriusCalculationPoint(final nl.overheid.aerius.shared.domain.v2.point.CIMLKCalculationPoint aeriusPoint) {
+    final CIMLKCalculationPoint point = new CIMLKCalculationPoint();
+    point.setRejectionGrounds(aeriusPoint.getRejectionGrounds());
+    point.setMonitorSubstance(aeriusPoint.getMonitorSubstance());
+    return point;
   }
 
   private AbstractCalculationPoint fromSubPoint(final nl.overheid.aerius.shared.domain.v2.point.SubPoint aeriusPoint) {
